Replaced old MPBase-based resources by new ones based on Resource<T>
