+++ px-dotnet/MercadoPagoSDK.csproj
@@ -95,24 +95,19 @@
     <Compile Include="Exceptions\MPConfException.cs" />
     <Compile Include="Exceptions\MPException.cs" />
     <Compile Include="Exceptions\MPRESTException.cs" />
-    <Compile Include="Resources\New\IdentificationType.cs" />
-    <Compile Include="Resources\New\Card.cs" />
-    <Compile Include="Resources\New\Customer.cs" />
-    <Compile Include="Resources\New\MerchantOrder.cs" />
-    <Compile Include="Resources\New\Payment.cs" />
-    <Compile Include="Resources\New\PaymentMethod.cs" />
-    <Compile Include="Resources\New\Plan.cs" />
-    <Compile Include="Resources\New\Preapproval.cs" />
-    <Compile Include="Resources\New\Preference.cs" />
+    <Compile Include="Resources\IdentificationType.cs" />
+    <Compile Include="Resources\Card.cs" />
+    <Compile Include="Resources\Customer.cs" />
+    <Compile Include="Resources\MerchantOrder.cs" />
+    <Compile Include="Resources\Payment.cs" />
+    <Compile Include="Resources\PaymentMethod.cs" />
+    <Compile Include="Resources\Plan.cs" />
+    <Compile Include="Resources\Preapproval.cs" />
+    <Compile Include="Resources\Preference.cs" />
     <Compile Include="SDK.cs" />
     <Compile Include="Net\MPRESTClient.cs" />
     <Compile Include="Net\MPRequest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Resources\Card.cs" />
-    <Compile Include="Resources\Customer.cs" />
-    <Compile Include="Resources\Payment.cs" />
-    <Compile Include="Resources\MerchantOrder.cs" />
-    <Compile Include="Resources\Preference.cs" />
     <Compile Include="DataStructures\Customer\CustomerAddress.cs" />
     <Compile Include="DataStructures\Customer\Card\CardHolder.cs" />
     <Compile Include="DataStructures\Customer\Card\Identification.cs" />
@@ -161,9 +156,7 @@
     <Compile Include="DataStructures\Preference\Phone.cs" />
     <Compile Include="DataStructures\Preference\ReceiverAddress.cs" />
     <Compile Include="DataStructures\Preference\Shipment.cs" />
-    <Compile Include="Resources\Plan.cs" />
     <Compile Include="Resources\Subscription.cs" />
-    <Compile Include="Resources\Preapproval.cs" />
     <Compile Include="DataStructures\Plan\Autorecurring.cs" />
     <Compile Include="DataStructures\Plan\FreeTrial.cs" />
     <Compile Include="Common\CurrencyId.cs" />
@@ -184,7 +177,6 @@
     <Compile Include="DataStructures\Payment\Barcode.cs" />
     <Compile Include="Common\EncodingType.cs" />
     <Compile Include="DataStructures\Payment\AdditionalInfoAddress.cs" />
-    <Compile Include="Resources\PaymentMethod.cs" />
     <Compile Include="DataStructures\PaymentMethod\Settings.cs" />
     <Compile Include="DataStructures\PaymentMethod\Bin.cs" />
     <Compile Include="DataStructures\PaymentMethod\CardNumber.cs" />
