+++ px-dotnet/Resources/Customer.cs
@@ -1,178 +1,136 @@
-﻿using MercadoPago.Resources;
-using MercadoPago.DataStructures.Customer;
-using Newtonsoft.Json.Linq;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
+using MercadoPago.DataStructures.Customer;
+using Newtonsoft.Json.Linq;
 
 namespace MercadoPago
-{    
-    public class Customer : MPBase
+{
+    public sealed class Customer : Resource<Customer>
     {
         #region Actions
+
         /// <summary>
         /// Get all customers acoording to specific filters
         /// </summary>
-        [GETEndpoint("/v1/customers/search")]
-        public static List<Customer> Search(Dictionary<string, string> filters)
-        {
-            return Search(filters, WITHOUT_CACHE);
-        }
-        /// <summary>
-        /// Get all customers acoording to specific filters
-        /// </summary>
-        [GETEndpoint("/v1/customers/search")]
-        public static List<Customer> Search(Dictionary<string, string> filters, bool useCache)
-        {
-            return (List<Customer>)ProcessMethodBulk<Customer>(typeof(Customer), "Search", filters, useCache);
-        } 
+        public static List<Customer> Search(Dictionary<string, string> filters, bool useCache = false) => 
+            GetList("/v1/customers/search", useCache, filters);
+
+        public static IQueryable<Customer> Query(bool useCache = false) =>
+            CreateQuery("/v1/customers/search", useCache);
+
         /// <summary>
         /// Find a customer by ID.
         /// </summary>
         /// <param name="id">Customer ID.</param>
         /// <param name="useCache">Cache configuration.</param>
         /// <returns>Searched customer.</returns>
-        [GETEndpoint("/v1/customers/:id")]
-        public static Customer FindById(string id, bool useCache)
-        {
-            return (Customer)ProcessMethod<Customer>("FindById", id, useCache);
-        }
-        /// <summary>
-        /// Find a customer by ID.
-        /// </summary>
-        [GETEndpoint("/v1/customers/:id")]
-        public static Customer FindById(String id)
-        {
-            return FindById(id, WITHOUT_CACHE);
-        }
+        public static Customer FindById(string id, bool useCache = false) => Get($"/v1/customers/{id}", useCache);
+
         /// <summary>
         /// Save a new customer
         /// </summary>
-        [POSTEndpoint("/v1/customers")]
-        public Customer Save()
-        {
-            return (Customer)ProcessMethod<Customer>("Save", WITHOUT_CACHE);
-        }
+        public Customer Save() => Post("/v1/customers");
+
         /// <summary>
         /// Update editable properties
         /// </summary>
-        [PUTEndpoint("/v1/customers/:id")]
-        public Customer Update()
-        {
-            return (Customer)ProcessMethod<Customer>("Update", WITHOUT_CACHE);
-        }
+        public Customer Update() => Get($"/v1/customers/{Id}");
+
         /// <summary>
         /// Remove a customer
         /// </summary>
-        [DELETEEndpoint("/v1/customers/:id")]
-        public Customer Delete()
-        {
-            return (Customer)ProcessMethod<Customer>("Delete", WITHOUT_CACHE);
-        } 
+        public Customer Delete() => Delete($"/v1/customers/{Id}");
+        
         #endregion
 
         #region Properties 
-        private string _id;
-        private string _email;
-        private string _first_name;
-        private string _last_name;
-        private Phone? _phone;
-        private Identification? _identification;
-        private string _default_address;
-        private Address? _address;
-        private DateTime? _date_registered;
-        private string _description;
-        private DateTime? _date_created;
-        private DateTime? _date_last_updated;
-        private JObject _metadata;
-        private string _default_card;
-        private List<Card> _cards = new List<Card>();
-        private List<CustomerAddress> _addresses = new List<CustomerAddress>();
-        private bool? _live_mode; 
-        #endregion
 
-        #region Accessors
         /// <summary>
         /// Customer ID
         /// </summary>
-        public string Id { 
-            get {  return _id; } 
-            set {  _id = value; } 
-        }
+        public string Id { get; set; }
+
         /// <summary>
         /// Customer's email
         /// </summary>
-        public string Email { 
-            get {  return _email; } 
-            set {  _email = value; } 
-        }
+        public string Email { get; set; }
+
         /// <summary>
         /// Customer's name
         /// </summary>
-        public string FirstName { 
-            get {  return _first_name; } 
-            set {  _first_name = value; } 
-        }
+        public string FirstName { get; set; }
+
         /// <summary>
         /// Customer's last name
         /// </summary>
-        public string LastName { 
-            get {  return _last_name; } 
-            set {  _last_name = value; } 
-        }
+        public string LastName { get; set; }
+
         /// <summary>
         /// Customer phone's information
         /// </summary>
-        public Phone? Phone { get {  return _phone; } set {  _phone = value; } }
+        public Phone? Phone { get; set; }
+
         /// <summary>
         /// Customer identification's information
         /// </summary>
-        public Identification? Identification { get {  return _identification; } set {  _identification = value; } }
+        public Identification? Identification { get; set; }
+
         /// <summary>
         /// Customer's default address
         /// </summary>
-        public string DefaultAddress { get {  return _default_address; } set {  _default_address = value; } }
+        public string DefaultAddress { get; set; }
+
         /// <summary>
         /// Default address's information
         /// </summary>
-        public Address? Address { get { return _address; } set {  _address = value; } }
+        public Address? Address { get; set; }
+
         /// <summary>
         /// Customer's registration date
         /// </summary>
-        public DateTime? DateRegistered { get { return _date_registered; } set {  _date_registered = value; } }
+        public DateTime? DateRegistered { get; set; }
+
         /// <summary>
         /// Customer's description
         /// </summary>
-        public string Description { get { return _description; } set {  _description = value; } }
+        public string Description { get; set; }
+
         /// <summary>
         /// Customer's date created
         /// </summary>
-        public DateTime? DateCreated { get { return _date_created; } set {  _date_created = value; } }
+        public DateTime? DateCreated { get; set; }
+
         /// <summary>
         /// Last modified date
         /// </summary>
-        public DateTime? DateLastUpdated { get { return _date_last_updated; } set {  _date_last_updated = value; } }
+        public DateTime? DateLastUpdated { get; set; }
+
         /// <summary>
         /// Customer's metadata
         /// </summary>
-        public JObject Metadata { get { return _metadata; } set {  _metadata = value; } }
+        public JObject Metadata { get; set; }
+
         /// <summary>
         /// Customer's default card
         /// </summary>
-        public string DefaultCard { get { return _default_card; } set {  _default_card = value; } }
+        public string DefaultCard { get; set; }
+
         /// <summary>
         /// Customer's cards
         /// </summary>
-        public List<Card> Cards { get { return _cards; } set {  _cards = value; } }
+        public List<Card> Cards { get; set; } = new List<Card>();
+
         /// <summary>
         /// Customer's addresses
         /// </summary>
-        public List<CustomerAddress> Addresses { get { return _addresses; } set {  _addresses = value; } }
+        public List<CustomerAddress> Addresses { get; set; } = new List<CustomerAddress>();
+
         /// <summary>
         /// Whether the customers will be in sandbox or in production mode
         /// </summary>
-        public bool? LiveMode { get { return _live_mode; } set {  _live_mode = value; } }
+        public bool? LiveMode { get; set; }
+
         #endregion
     }
 }
