+++ px-dotnet/Resources/MerchantOrder.cs
@@ -1,218 +1,90 @@
-﻿using MercadoPago.DataStructures.MerchantOrder;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
+using MercadoPago.DataStructures.MerchantOrder;
 
-namespace MercadoPago.Resources
+namespace MercadoPago
 {
-    public class MerchantOrder : MPBase
+    public sealed class MerchantOrder : Resource<MerchantOrder>
     {
         #region Actions
-        
-        public MerchantOrder Load(string id) 
-        {
-            return Load(id, WITHOUT_CACHE);
-        }
 
-        [GETEndpoint("/merchant_order/:id")]
-        public MerchantOrder Load(string id, bool useCache)
-        {
-            return (MerchantOrder)ProcessMethod<MerchantOrder>(typeof(MerchantOrder), "Load", id, useCache);
-        }
-        
-        [POSTEndpoint("/merchant_order")]
-        public MerchantOrder Save() 
-        {
-            return (MerchantOrder)ProcessMethod<MerchantOrder>("Save", WITHOUT_CACHE);
-        }
-        
-        [PUTEndpoint("/merchant_order/:id")]
-        public MerchantOrder Update() 
-        {
-            return (MerchantOrder)ProcessMethod<MerchantOrder>("Update", WITHOUT_CACHE);
-        }
+        public MerchantOrder Load(string id, bool useCache = false) => Get($"/merchant_order/{id}", useCache);
+
+        public MerchantOrder Save() => Post("/merchant_order");
+
+        public MerchantOrder Update() => Put($"/merchant_order/{Id}");
 
         #endregion
 
         #region Properties
 
-        private string id;
-        private string preferenceId;
-        private DateTime? dateCreated;
-        private DateTime? lastUpdate;
-        private string applicationId;
-        private string status;
-        private string siteId;
-        private Payer payer;
-        private Collector collector;
-        private int sponsorId;
-        private List<MerchantOrderPayment> payments;
-        private decimal? paidAmount;
-        private decimal? refundedAmount;
-        private decimal? shippingCost;
-        private bool? cancelled;
-        private List<Item> items;
-        private List<Shipment> shipments;
+        public string Id { get; set; }
+
+        public string PreferenceId { get; set; }
+
+        public DateTime? DateCreated { get; }
+
+        public DateTime? LastUpdate { get; }
+
+        public string ApplicationId { get; set; }
+
+        public string Status { get; }
+
+        public string SiteId { get; set; }
+
+        public Payer Payer { get; set; }
+
+        public Collector Collector { get; set; }
+
+        public int SponsorId { get; set; }
+
+        public List<MerchantOrderPayment> Payments { get; }
+
+        public decimal? PaidAmount { get; }
+
+        public decimal? RefundedAmount { get; }
+
+        public decimal? ShippingCost { get; }
+
+        public bool? Cancelled { get; set; }
+
+        public List<Item> Items { get; set; }
+
+        public List<Shipment> Shipments { get; set; }
+
         [StringLength(500)]
-        private string notificationUrl;
+        public string NotificationUrl { get; set; }
+
         [StringLength(600)]
-        private string additionalInfo;
+        public string AdditionalInfo { get; set; }
+
         [StringLength(256)]
-        private string externalReference;
+        public string ExternalReference { get; set; }
+
         [StringLength(256)]
-        private string marketplace;
-        private decimal? totalAmount;
+        public string Marketplace { get; set; }
+
+        public decimal? TotalAmount { get; }
 
         #endregion
 
-        #region Accessors
-       
-        public string ID
+        public void AppendShipment(Shipment shipment)
         {
-            get { return id; }
-            set { this.id = value; } //This Accessor must be removed after testing approvement.
-        }
-       
-        public string PreferenceId
-        {
-            get { return preferenceId; }
-            set { preferenceId = value; }
-        }        
-
-        public DateTime? DateCreated
-        {
-            get { return dateCreated; }            
-        }
-
-        
-
-        public DateTime? LastUpdate
-        {
-            get { return lastUpdate; }            
-        }
-
-        public string ApplicationId
-        {
-            get { return applicationId; }
-            set { applicationId = value; }
-        }
-
-        public string Status
-        {
-            get { return status; }            
-        }
-
-        public string SiteId
-        {
-            get { return siteId; }
-            set { siteId = value; }
-        }
-
-        public Payer Payer
-        {
-            get { return payer; }
-            set { payer = value; }
-        }
-
-        public Collector Collector
-        {
-            get { return collector; }
-            set { collector = value; }
-        }
-
-        public int SponsorId
-        {
-            get { return sponsorId; }
-            set { sponsorId = value; }
-        }
-
-        public List<MerchantOrderPayment> Payments
-        {
-            get { return payments; }            
-        }        
-
-        public decimal? PaidAmount
-        {
-            get { return paidAmount; }            
-        }
-       
-        public decimal? RefundedAmount
-        {
-            get { return refundedAmount; }            
-        }
-
-        public decimal? ShippingCost
-        {
-            get { return shippingCost; }            
-        }
-
-        public bool? Cancelled
-        {
-            get { return cancelled; }
-            set { cancelled = value; }
-        }
-
-        public List<Item> Items
-        {
-            get { return items; }
-            set { items = value; }
+            if (Shipments == null)
+            {
+                Shipments = new List<Shipment>();
+            }
+            Shipments.Add(shipment);
         }
 
         public void AppendItem(Item item)
         {
-            if (items == null)
+            if (Items == null)
             {
-                items = new List<Item>();
+                Items = new List<Item>();
             }
-            items.Add(item);            
-        }        
-
-        public List<Shipment> Shipments
-        {
-            get { return shipments; }
-            set { shipments = value; }
+            Items.Add(item);
         }
-
-        public void AppendShipment(Shipment shipment)
-        {
-            if (shipments == null)
-            {
-                shipments = new List<Shipment>();
-            }
-            shipments.Add(shipment);            
-        }
-
-        public string NotificationUrl
-        {
-            get { return notificationUrl; }
-            set { notificationUrl = value; }
-        }
-
-        public string AdditionalInfo
-        {
-            get { return additionalInfo; }
-            set { additionalInfo = value; }
-        }
-
-        public string ExternalReference
-        {
-            get { return externalReference; }
-            set { externalReference = value; }
-        }
-
-        public string Marketplace
-        {
-            get { return marketplace; }
-            set { marketplace = value; }
-        }
-       
-        public decimal? TotalAmount
-        {
-            get { return totalAmount; }
-        }
-
-        #endregion
     }
 }
